Update evaluation scripts
