--- evaluation/scripts/EXACT_COSTS_UPDATE.md
@@ -1,156 +1,0 @@
-# ✅ EXACT TogoMCP Cost Tracking Now Available!
-
-## Great News! 🎉
-
-The Agent SDK **DOES** expose token usage information in the `ResultMessage.usage` field. This means we can now get **exact** TogoMCP costs instead of estimates!
-
-## What Was Updated
-
-### 1. `automated_test_runner.py`
-**Added token tracking from Agent SDK:**
-```python
-if isinstance(message, ResultMessage):
-    if hasattr(message, 'usage'):
-        usage_info = message.usage
-```
-
-**New CSV fields:**
-- `togomcp_input_tokens` - Exact input token count
-- `togomcp_output_tokens` - Exact output token count
-- `togomcp_cache_creation_input_tokens` - Tokens used for prompt caching creation
-- `togomcp_cache_read_input_tokens` - Tokens read from cache
-
-### 2. `compute_costs.py`
-**Smart cost calculation:**
-- Automatically detects if actual token counts are available
-- Uses exact tokens when available
-- Falls back to estimation for old CSVs without token data
-- Shows cache usage statistics
-
-## Usage Information from Agent SDK
-
-From your test run, the Agent SDK provides:
-```python
-usage: {
-    'input_tokens': 17,                      # Regular input tokens
-    'cache_creation_input_tokens': 17474,    # Tokens for creating prompt cache
-    'cache_read_input_tokens': 34565,        # Tokens read from cache
-    'output_tokens': 490,                    # Output tokens
-    'server_tool_use': {
-        'web_search_requests': 0,
-        'web_fetch_requests': 0
-    },
-    'service_tier': 'standard',
-    'cache_creation': {
-        'ephemeral_1h_input_tokens': 0,
-        'ephemeral_5m_input_tokens': 17474
-    }
-}
-```
-
-## How to Use
-
-### Run a new evaluation (captures token data):
-```bash
-python automated_test_runner.py questions.json
-```
-
-### Compute costs (automatically uses exact tokens):
-```bash
-python compute_costs.py evaluation_results.csv
-```
-
-You'll see:
-```
-TOGOMCP COSTS (With MCP Tools) (EXACT)
---------------------------------------------------------------------------------
-  ✅ Using actual token counts from Agent SDK ResultMessage
-
-  Successful tests:     10/10
-  Input tokens:         17
-  Output tokens:        490
-  Cache creation:       17,474 tokens
-  Cache read:           34,565 tokens
-  Total tokens:         507
-  Total cost:           $0.0074
-```
-
-## Cost Calculation
-
-The costs are now calculated from actual token counts:
-- **Input tokens**: $3.00 per million tokens
-- **Output tokens**: $15.00 per million tokens
-- **Cache read**: Typically cheaper (90% discount), but for now counted as regular input
-
-## Prompt Caching Impact
-
-The large `cache_creation_input_tokens` (17,474) and `cache_read_input_tokens` (34,565) show that:
-- MCP tool schemas are being cached (good for efficiency!)
-- Cache reads significantly reduce costs over multiple requests
-- The Agent SDK is automatically using prompt caching
-
-## Backward Compatibility
-
-The system works with both:
-- ✅ **New CSVs**: Use exact token counts from Agent SDK
-- ✅ **Old CSVs**: Fall back to estimation method
-
-## Comparison: Estimated vs Actual
-
-### Your Previous Run (Estimated):
-```
-TogoMCP: $0.3129 (estimated)
-Overhead: +1117.3% vs baseline
-```
-
-### New Run with Actual Tokens:
-The actual costs will likely be lower due to:
-1. Exact token counts (no 4x character approximation)
-2. Prompt caching reducing repeat costs
-3. More accurate tool overhead calculation
-
-## Summary
-
-| Aspect | Before | After |
-|--------|--------|-------|
-| **TogoMCP Costs** | ❌ Estimated (~30% error) | ✅ Exact from API |
-| **Method** | Text length ÷ 4 | Actual token counts |
-| **Cache Info** | ❌ Not available | ✅ Shows creation & reads |
-| **Accuracy** | ~±30% | 100% accurate |
-| **Requires** | Nothing | Re-run evaluation |
-
-## Next Steps
-
-1. **Re-run your evaluation** to capture token data:
-   ```bash
-   python automated_test_runner.py example_questions.json
-   ```
-
-2. **Compute exact costs**:
-   ```bash
-   python compute_costs.py evaluation_results.csv
-   ```
-
-3. **Compare** with your previous estimated costs
-
-4. **Update budgets** based on actual costs
-
-## Files Modified
-
-1. ✅ `/Users/arkinjo/work/GitHub/togo-mcp/evaluation/scripts/automated_test_runner.py`
-   - Captures token usage from ResultMessage
-   - Adds token fields to CSV output
-
-2. ✅ `/Users/arkinjo/work/GitHub/togo-mcp/evaluation/scripts/compute_costs.py`
-   - Uses actual tokens when available
-   - Falls back to estimation for old CSVs
-   - Shows cache statistics
-
-3. ✅ Documentation files created:
-   - `GETTING_ACTUAL_COSTS.md`
-   - `inspect_agent_sdk_tokens.py`
-   - `query_anthropic_usage.py`
-
-## Questions?
-
-The Agent SDK's ResultMessage provides comprehensive usage information, so you now have 100% accurate cost tracking for both baseline and TogoMCP evaluations!
