+++ evaluation/scripts/quick_start.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
-# Quick Start Script for TogoMCP Evaluation with Agent SDK
+# Quick Start Script for TogoMCP Evaluation
 
 set -e
 
 echo "============================================="
-echo "TogoMCP Evaluation - Agent SDK Quick Start"
+echo "TogoMCP Evaluation - Quick Start"
 echo "============================================="
 echo ""
 
@@ -13,12 +13,12 @@ if [ -z "$ANTHROPIC_API_KEY" ]; then
     echo "❌ Error: ANTHROPIC_API_KEY not set"
     echo ""
     echo "Please set your API key:"
-    echo "  export ANTHROPIC_API_KEY='your-api-key-here'"
+    echo "  export ANTHROPIC_API_KEY='sk-ant-your-key-here'"
     echo ""
     exit 1
 fi
 
-echo "✓ API key found"
+echo "✅ API key found"
 
 # Check if claude-agent-sdk is installed
 if ! python3 -c "import claude_agent_sdk" 2>/dev/null; then
@@ -26,7 +26,7 @@ if ! python3 -c "import claude_agent_sdk" 2>/dev/null; then
     echo "📦 Installing required packages..."
     pip install -r requirements.txt
 else
-    echo "✓ claude-agent-sdk installed"
+    echo "✅ claude-agent-sdk installed"
 fi
 
 # Check if anthropic is installed
@@ -34,11 +34,23 @@ if ! python3 -c "import anthropic" 2>/dev/null; then
     echo "📦 Installing anthropic..."
     pip install anthropic
 else
-    echo "✓ anthropic installed"
+    echo "✅ anthropic installed"
+fi
+
+# Check if pandas is installed (for compute_costs.py)
+if ! python3 -c "import pandas" 2>/dev/null; then
+    echo "📦 Installing pandas..."
+    pip install pandas
+else
+    echo "✅ pandas installed"
 fi
 
 echo ""
-echo "Running evaluation with example questions..."
+echo "============================================="
+echo "Step 1: Running Evaluation"
+echo "============================================="
+echo ""
+echo "Testing with example questions..."
 echo ""
 
 # Run the test runner
@@ -46,18 +58,42 @@ python3 automated_test_runner.py example_questions.json
 
 echo ""
 echo "============================================="
-echo "Evaluation Complete!"
+echo "Step 2: Calculating Costs"
 echo "============================================="
 echo ""
-echo "Results saved to: evaluation_results.csv"
+
+# Calculate costs
+python3 compute_costs.py evaluation_results.csv
+
+echo ""
+echo "============================================="
+echo "Step 3: Analyzing Results"
+echo "============================================="
+echo ""
+
+# Analyze results
+python3 results_analyzer.py evaluation_results.csv
+
+echo ""
+echo "============================================="
+echo "✅ Evaluation Complete!"
+echo "============================================="
+echo ""
+echo "📊 Results saved to: evaluation_results.csv"
 echo ""
 echo "Next steps:"
-echo "  1. Review: open evaluation_results.csv"
-echo "  2. Check tools_used column for MCP tool names"
-echo "  3. Compare baseline_text vs togomcp_text"
+echo "  1. Review results: open evaluation_results.csv"
+echo "  2. Check cost breakdown above"
+echo "  3. Generate dashboard: python3 generate_dashboard.py evaluation_results.csv --open"
 echo "  4. Create your own questions: cp example_questions.json my_questions.json"
 echo ""
 echo "Documentation:"
-echo "  - README.md - Full usage guide"
-echo "  - MCP_CONFIGURATION.md - MCP server setup"
+echo "  - README.md - Complete usage guide with cost analysis"
+echo "  - QUICK_REFERENCE.md - Quick command reference"
+echo "  - QUESTION_FORMAT.md - Question file format"
+echo ""
+echo "Cost information:"
+echo "  - Per question: ~\$0.055 (\$0.002 baseline + \$0.053 TogoMCP)"
+echo "  - Uses isolated sessions for optimal cache efficiency"
+echo "  - Cache costs are stable and predictable"
 echo ""
