--- evaluation/scripts/test_3questions.json
@@ -1,20 +1,0 @@
-[
-  {
-    "id": 1,
-    "question": "What is the UniProt ID for SpCas9 from Streptococcus pyogenes?",
-    "category": "Precision",
-    "expected_answer": "Q99ZW2"
-  },
-  {
-    "id": 2,
-    "question": "What is the PubChem CID for caffeine?",
-    "category": "Precision",
-    "expected_answer": "2519"
-  },
-  {
-    "id": 3,
-    "question": "What is the ChEMBL ID for aspirin?",
-    "category": "Precision",
-    "expected_answer": "CHEMBL25"
-  }
-]
