--- evaluation/scripts/test_results.csv
@@ -1,16 +1,0 @@
-question_id,date,category,question_text,baseline_success,baseline_actually_answered,baseline_has_expected,baseline_confidence,baseline_text,baseline_error,baseline_time,baseline_input_tokens,baseline_output_tokens,togomcp_success,togomcp_has_expected,togomcp_confidence,togomcp_text,togomcp_error,togomcp_time,togomcp_input_tokens,togomcp_output_tokens,togomcp_cache_creation_input_tokens,togomcp_cache_read_input_tokens,tools_used,tool_details,value_add,expected_answer,notes
-1,2025-12-18,Precision,What is the UniProt ID for SpCas9 from Streptococcus pyogenes?,True,True,True,1.0,"The UniProt ID for SpCas9 (Cas9 from Streptococcus pyogenes) is **Q99ZW2**.
-
-This entry corresponds to the CRISPR-associated endonuclease Cas9 from Streptococcus pyogenes serotype M1, which is the most commonly used Cas9 variant in CRISPR-Cas9 gene editing applications.",,3.3675639629364014,61,99,True,True,1.0,"The UniProt ID for SpCas9 from Streptococcus pyogenes is **Q99ZW2**.
-
-This entry corresponds to the CRISPR-associated endonuclease Cas9/Csn1 (EC 3.1.-.-), also known as SpCas9 or SpyCas9, from Streptococcus pyogenes serotype M1. This is the widely used Cas9 protein that has become the standard tool for CRISPR gene editing applications.",,10.436583995819092,9,242,211,32688,mcp__togomcp__search_uniprot_entity,"[{""name"": ""mcp__togomcp__search_uniprot_entity"", ""input"": {""query"": ""SpCas9 Streptococcus pyogenes"", ""limit"": 10}}]",VALUABLE,Q99ZW2,
-2,2025-12-18,Precision,What is the PubChem CID for caffeine?,True,True,True,1.0,"The PubChem CID (Compound Identifier) for caffeine is 2519.
-
-Caffeine's chemical name is 1,3,7-trimethylxanthine, and it has the molecular formula C8H10N4O2. PubChem CID 2519 is the standard identifier used in the PubChem database to reference caffeine's chemical structure and properties.",,3.7137837409973145,50,93,True,True,1.0,"The PubChem CID for caffeine is **2519**.
-
-This is the unique identifier used by PubChem to reference caffeine in their chemical database. You can access the full compound information for caffeine at: https://pubchem.ncbi.nlm.nih.gov/compound/2519",,10.33234715461731,9,163,0,32807,mcp__togomcp__get_pubchem_compound_id,"[{""name"": ""mcp__togomcp__get_pubchem_compound_id"", ""input"": {""compound_name"": ""caffeine""}}]",VALUABLE,2519,
-3,2025-12-18,Precision,What is the ChEMBL ID for aspirin?,True,True,True,1.0,"The ChEMBL ID for aspirin is CHEMBL25.
-
-Aspirin (acetylsalicylic acid) is a well-known compound in the ChEMBL database, which is a large-scale bioactivity database maintained by the European Bioinformatics Institute (EBI).",,3.870850086212158,50,72,True,True,1.0,"Based on the search results, the ChEMBL ID for aspirin is **CHEMBL25**. This is the top result with the highest score (36.594707) and is listed as a compound entity type, which matches what we're looking for.
-
-CHEMBL25 represents aspirin (acetylsalicylic acid) in the ChEMBL database, which is a large-scale bioactivity database that contains information about drug-like bioactive compounds and their activities against biological targets.",,11.150093078613281,9,206,0,33319,mcp__togomcp__search_chembl_id_lookup,"[{""name"": ""mcp__togomcp__search_chembl_id_lookup"", ""input"": {""query"": ""aspirin""}}]",VALUABLE,CHEMBL25,
