--- evaluation/scripts/test_results_independent.csv
@@ -1,19 +1,0 @@
-question_id,date,category,question_text,baseline_success,baseline_actually_answered,baseline_has_expected,baseline_confidence,baseline_text,baseline_error,baseline_time,baseline_input_tokens,baseline_output_tokens,togomcp_success,togomcp_has_expected,togomcp_confidence,togomcp_text,togomcp_error,togomcp_time,togomcp_input_tokens,togomcp_output_tokens,togomcp_cache_creation_input_tokens,togomcp_cache_read_input_tokens,tools_used,tool_details,value_add,expected_answer,notes
-1,2025-12-18,Precision,What is the UniProt ID for SpCas9 from Streptococcus pyogenes?,True,True,True,1.0,"The UniProt ID for SpCas9 (Cas9 from Streptococcus pyogenes) is **Q99ZW2**.
-
-This corresponds to the Cas9 protein from Streptococcus pyogenes serotype M1, which is the most commonly used Cas9 variant in CRISPR-Cas9 genome editing applications.",,4.250576972961426,61,88,True,True,1.0,"Perfect! The UniProt ID for SpCas9 (Cas9 from Streptococcus pyogenes) is **Q99ZW2**.
-
-This entry corresponds to:
-- **Protein name**: CRISPR-associated endonuclease Cas9/Csn1 (EC 3.1.-.-) (SpCas9) (SpyCas9)
-- **Organism**: Streptococcus pyogenes serotype M1
-- **UniProt ID**: Q99ZW2
-
-This is the most commonly referenced and well-characterized Cas9 protein used in CRISPR-Cas9 genome editing applications. The Q99ZW2 entry represents the canonical SpCas9 protein that has been extensively studied and is widely used in research and biotechnology applications.",,11.359058856964111,9,329,17061,16344,mcp__togomcp__search_uniprot_entity,"[{""name"": ""mcp__togomcp__search_uniprot_entity"", ""input"": {""query"": ""Cas9 Streptococcus pyogenes"", ""limit"": 10}}]",VALUABLE,Q99ZW2,
-2,2025-12-18,Precision,What is the PubChem CID for caffeine?,True,True,True,1.0,The PubChem CID (Compound Identifier) for caffeine is 2519.,,14.468032836914062,50,24,True,True,1.0,"The PubChem CID for caffeine is **2519**.
-
-This is the unique identifier used in the PubChem database to reference caffeine (1,3,7-trimethylxanthine), which is the well-known stimulant compound found in coffee, tea, and many other beverages and foods.",,8.345113039016724,9,165,291,32516,mcp__togomcp__get_pubchem_compound_id,"[{""name"": ""mcp__togomcp__get_pubchem_compound_id"", ""input"": {""compound_name"": ""caffeine""}}]",VALUABLE,2519,
-3,2025-12-18,Precision,What is the ChEMBL ID for aspirin?,True,True,True,1.0,"The ChEMBL ID for aspirin is CHEMBL25.
-
-Aspirin (acetylsalicylic acid) is a well-known anti-inflammatory drug and analgesic, and CHEMBL25 is its identifier in the ChEMBL database, which is a large-scale bioactivity database maintained by the European Bioinformatics Institute (EBI).",,4.9660258293151855,50,89,True,True,1.0,"The ChEMBL ID for aspirin is **CHEMBL25**. This is the top result with the highest score (36.59) and is identified as a compound entity type, which matches what we're looking for.
-
-CHEMBL25 corresponds to aspirin (acetylsalicylic acid), which is one of the most well-known and widely used pharmaceutical compounds in the ChEMBL database.",,9.67300295829773,9,187,803,32516,mcp__togomcp__search_chembl_id_lookup,"[{""name"": ""mcp__togomcp__search_chembl_id_lookup"", ""input"": {""query"": ""aspirin""}}]",VALUABLE,CHEMBL25,
